+++ sdk/trafficmanager/azure-mgmt-trafficmanager/azure/mgmt/trafficmanager/models/_traffic_manager_management_client_enums.py
@@ -7,22 +7,29 @@
 # --------------------------------------------------------------------------
 
 from enum import Enum
-from six import with_metaclass
 from azure.core import CaseInsensitiveEnumMeta
 
 
-class AllowedEndpointRecordType(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
-    """The allowed type DNS record types for this profile.
-    """
+class AllowedEndpointRecordType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The allowed type DNS record types for this profile."""
 
     DOMAIN_NAME = "DomainName"
     I_PV4_ADDRESS = "IPv4Address"
     I_PV6_ADDRESS = "IPv6Address"
     ANY = "Any"
 
-class EndpointMonitorStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
-    """The monitoring status of the endpoint.
+
+class AlwaysServe(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """If Always Serve is enabled, probing for endpoint health will be disabled and endpoints will be
+    included in the traffic routing method.
     """
+
+    ENABLED = "Enabled"
+    DISABLED = "Disabled"
+
+
+class EndpointMonitorStatus(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The monitoring status of the endpoint."""
 
     CHECKING_ENDPOINT = "CheckingEndpoint"
     ONLINE = "Online"
@@ -31,7 +38,8 @@ class EndpointMonitorStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     INACTIVE = "Inactive"
     STOPPED = "Stopped"
 
-class EndpointStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
+
+class EndpointStatus(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """The status of the endpoint. If the endpoint is Enabled, it is probed for endpoint health and is
     included in the traffic routing method.
     """
@@ -39,23 +47,25 @@ class EndpointStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     ENABLED = "Enabled"
     DISABLED = "Disabled"
 
-class EndpointType(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
+
+class EndpointType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """EndpointType."""
 
     AZURE_ENDPOINTS = "AzureEndpoints"
     EXTERNAL_ENDPOINTS = "ExternalEndpoints"
     NESTED_ENDPOINTS = "NestedEndpoints"
 
-class MonitorProtocol(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
-    """The protocol (HTTP, HTTPS or TCP) used to probe for endpoint health.
-    """
+
+class MonitorProtocol(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The protocol (HTTP, HTTPS or TCP) used to probe for endpoint health."""
 
     HTTP = "HTTP"
     HTTPS = "HTTPS"
     TCP = "TCP"
 
-class ProfileMonitorStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
-    """The profile-level monitoring status of the Traffic Manager profile.
-    """
+
+class ProfileMonitorStatus(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The profile-level monitoring status of the Traffic Manager profile."""
 
     CHECKING_ENDPOINTS = "CheckingEndpoints"
     ONLINE = "Online"
@@ -63,16 +73,16 @@ class ProfileMonitorStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     DISABLED = "Disabled"
     INACTIVE = "Inactive"
 
-class ProfileStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
-    """The status of the Traffic Manager profile.
-    """
+
+class ProfileStatus(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The status of the Traffic Manager profile."""
 
     ENABLED = "Enabled"
     DISABLED = "Disabled"
 
-class TrafficRoutingMethod(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
-    """The traffic routing method of the Traffic Manager profile.
-    """
+
+class TrafficRoutingMethod(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The traffic routing method of the Traffic Manager profile."""
 
     PERFORMANCE = "Performance"
     PRIORITY = "Priority"
@@ -81,7 +91,8 @@ class TrafficRoutingMethod(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     MULTI_VALUE = "MultiValue"
     SUBNET = "Subnet"
 
-class TrafficViewEnrollmentStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
+
+class TrafficViewEnrollmentStatus(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Indicates whether Traffic View is 'Enabled' or 'Disabled' for the Traffic Manager profile.
     Null, indicates 'Disabled'. Enabling this feature will increase the cost of the Traffic Manage
     profile.