+++ sdk/confidentialledger/azure-mgmt-confidentialledger/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 2.0.0b1 (2022-11-25)
+
+### Features Added
+
+  - Model LedgerProperties has a new parameter ledger_storage_account
+
+### Breaking Changes
+
+  - Removed operation group ConfidentialLedgerOperationsMixin
+
 ## 1.0.0 (2022-05-30)
 
 **Features**